nsec should be long (was getting cut off)

+++ src/util.h
@@ -3,9 +3,9 @@
 
 
 
-void rh_nsleep(unsigned int delay_ns);
-void rh_usleep(unsigned int delay_us);
-void rh_msleep(unsigned int delay_us);
+void rh_nsleep(long delay_ns);
+void rh_usleep(long delay_us);
+void rh_msleep(long delay_us);
